Moved to masm syntax in length

+++ lab_08/inc/funcs.h
@@ -1,7 +1,9 @@
 #ifndef __CHECK_FUNCS_H__
 #define __CHECK_FUNCS_H__
 
-int asmstrlen(const char *str);
+#include <stdlib.h>
+
+size_t asmstrlen(const char *str);
 
 void strcopy(char *dest, char *src, int len);
